requirements.txt
requests
toml
aiohttp
discord.py
python-dotenv